fix: add adapter class
