+++ projects/bin/train_model.py
@@ -4,6 +4,7 @@ import os
 import yaml
 import argparse
 from semantic_selector.model.one_to_one import NNFullyConnectedModel
+from semantic_selector.adapter.one_to_one import MySQLToNNFullyConnectAdapter
 from semantic_selector.mysql import InputTable
 
 
@@ -21,17 +22,16 @@ def main():
     parser.add_argument('--debug', action='store_true')
     args = parser.parse_args()
 
-    (training, tests) = InputTable(args.threashold).fetch_data(args.ratio_test, args.seed)
-
     model_name = args.model_name
     print("model type: %s" % (model_name))
     if model_name == "nn_fc":
         model = NNFullyConnectedModel()
+        adapter = MySQLToNNFullyConnectAdapter(args.threashold, args.ratio_test, args.seed)
+        model.train(adapter, args.epochs)
     else:
         print("model %s unknown" % (model_name))
         sys.exit(1)
 
-    model.train(training, tests, args.epochs)
     if args.debug:
         print("failing inferences\n")
         print("estimated, correct")
@@ -41,22 +41,6 @@ def main():
             if estimated_topic != correct_topic:
                 print(estimated_topic + "," + correct_topic)
 
-
-    print()
-
-    num_unknown = 0
-    for t in training:
-        if t.canonical_topic == 'unknown':
-            num_unknown += 1
-    print('# of unknown in training: ' + str(num_unknown))
-
-    num_unknown = 0
-    for t in tests:
-        if t.canonical_topic == 'unknown':
-            num_unknown += 1
-    print('# of unknown in test: ' + str(num_unknown))
-    print("# of test data: " + str(len(tests)))
-    print("# of training_data: " + str(len(training)))
     model.save()
 
 if __name__ == '__main__':
